Solved
minimal 2-3 tree-like data structure is good enough, no need to self balance. Test case 7 just involved a sorted or reverse-sorted array as input, which would cause a stack overflow with this 2-3 tree. Simple check if the input array is sorted/reverse-sorted, if so then output an easily computed answer. BOOM